+++ src/search/alphabeta.rs
@@ -2,7 +2,7 @@
 
 use std::cmp::max;
 use std::sync::atomic::Ordering::Relaxed;
-use std::sync::atomic::{AtomicBool, AtomicI64};
+use std::sync::atomic::{AtomicBool, AtomicI32};
 use std::sync::Mutex;
 use std::time::Instant;
 
@@ -19,9 +19,9 @@ use super::super::logic::movegen::available_player_actions;
 use super::eval::{evaluate_action, evaluate_action_terminal, evaluate_position_with_details};
 
 /// Starting beta value for the alphabeta search (starting alpha is equal to -beta)
-pub const BASE_BETA: i64 = 262_144;
+pub const BASE_BETA: i32 = 262_144;
 /// Starting alpha value for the alphabeta search (starting alpha is equal to -beta)
-pub const BASE_ALPHA: i64 = -BASE_BETA;
+pub const BASE_ALPHA: i32 = -BASE_BETA;
 
 #[cfg(feature = "nps-count")]
 use std::sync::atomic::AtomicU64;
@@ -40,9 +40,9 @@ pub fn search(
     current_player: Player,
     depth: u64,
     end_time: Option<Instant>,
-    scores: &Option<Vec<i64>>,
+    scores: &Option<Vec<i32>>,
     transposition_table: Option<&Mutex<SearchTable>>,
-) -> Option<(Action, i64, Vec<i64>)> {
+) -> Option<(Action, i32, Vec<i32>)> {
     if depth == 0 {
         return None;
     }
@@ -66,7 +66,7 @@ pub fn search(
         return None;
     }
 
-    let scores: Vec<i64> = if depth == 1 {
+    let scores: Vec<i32> = if depth == 1 {
         #[cfg(feature = "nps-count")]
         unsafe {
             increment_node_count(n_actions as u64);
@@ -93,7 +93,7 @@ pub fn search(
         let alpha = BASE_ALPHA;
         let beta = BASE_BETA;
 
-        let mut scores: Vec<i64> = vec![0; n_actions];
+        let mut scores: Vec<i32> = vec![0; n_actions];
         let first_eval = -evaluate_action(
             cells,
             1 - current_player,
@@ -106,7 +106,7 @@ pub fn search(
         );
         scores[0] = first_eval;
 
-        let alpha: AtomicI64 = AtomicI64::new(max(alpha, first_eval));
+        let alpha: AtomicI32 = AtomicI32::new(max(alpha, first_eval));
         // This will stop iteration if there is a cutoff
         let cut: AtomicBool = AtomicBool::new(alpha.load(Relaxed) > beta);
 
@@ -120,7 +120,7 @@ pub fn search(
                 let action = available_actions[order[k]];
                 *score = {
                     if cut.load(Relaxed) {
-                        i64::MIN
+                        i32::MIN
                     } else {
                         let eval = {
                             // Search with a null window
@@ -170,7 +170,7 @@ pub fn search(
         }
     }
 
-    let scores: Vec<i64> = reverse_argsort(&scores, &order);
+    let scores: Vec<i32> = reverse_argsort(&scores, &order);
 
     scores
         .iter()
@@ -192,9 +192,9 @@ pub fn search_iterative(
     end_time: Option<Instant>,
     verbose: bool,
     transposition_table: Option<&Mutex<SearchTable>>,
-) -> Option<(Action, i64)> {
-    let mut best_result: Option<(Action, i64)> = None;
-    let mut last_scores: Option<Vec<i64>> = None;
+) -> Option<(Action, i32)> {
+    let mut best_result: Option<(Action, i32)> = None;
+    let mut last_scores: Option<Vec<i32>> = None;
     let start_time = Instant::now();
     for depth in 1..=max_depth {
         if let Some(end_time) = end_time {
